+++ othellia.py
@@ -1,7 +1,6 @@
 import pygame
 
-from settings import cell_values
-from settings.cell_values import BLACK_VALUE, WHITE_VALUE
+from settings import values
 from settings.colors import BOARD_COLOR
 from settings.graphics import HEIGHT, WIDTH
 from src.game import Game
@@ -16,8 +15,8 @@ if __name__ == "__main__":
     piece_layout = PieceLayout()
     indicator_layout = IndicatorLayout()
 
-    endgame_message_black_won = EndgameMessage(BLACK_VALUE)
-    endgame_message_white_won = EndgameMessage(WHITE_VALUE)
+    endgame_message_black_won = EndgameMessage(values.BLACK_VALUE)
+    endgame_message_white_won = EndgameMessage(values.WHITE_VALUE)
     endgame_message_draw = EndgameMessage(None)
 
     game = Game()
@@ -54,9 +53,9 @@ if __name__ == "__main__":
 
         if game.is_over:
             match game.get_winner():
-                case cell_values.BLACK_VALUE:
+                case values.BLACK_VALUE:
                     endgame_message_black_won.draw(screen)
-                case cell_values.WHITE_VALUE:
+                case values.WHITE_VALUE:
                     endgame_message_white_won.draw(screen)
                 case _:
                     endgame_message_draw.draw(screen)
