+++ tests/test_static_evaluation.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from game import Game
-from settings.cell_values import BLACK_VALUE, WHITE_VALUE
+from settings.values import BLACK_VALUE, WHITE_VALUE
 from static_evaluation import StaticEvaluation
